+++ src/MarketDataCollector.Contracts/Domain/Events/MarketEvent.cs
@@ -13,10 +13,10 @@ public sealed record MarketEvent(
     MarketEventPayload? Payload,
     long Sequence = 0,
     string Source = "IB",
-    int SchemaVersion = 1,
+    byte SchemaVersion = 1,
     MarketEventTier Tier = MarketEventTier.Raw,
     string? CanonicalSymbol = null,
-    int CanonicalizationVersion = 0,
+    byte CanonicalizationVersion = 0,
     string? CanonicalVenue = null
 )
 {
